Add signup template,tests,views,paths for supplier
in this commit i will be adding the login functionality with the signup
and i will do the authenticate part for the supplier in order to continue
doing the project logic.

Fix product templates

In this i am fixing the product template as said
and all templates should be in the base DIR/template.

Add account field in supplier and user

in this commit i have changed many things in order to
get a correct project and to have the authentication ability
and future functions that is built in django user model
here are few things:
- I have added a one to one field for each user and supplier
that is pointed to a user account in django user model
- I have changed the migration tests in order to get
a correct test that do not fail
- Updated some of few functions in client and supplier
- Update confest file in order to get rid off warnings
- Updated pytest test files to have a correct tests

nothing

Add account field in supplier and user

in this commit i have changed many things in order to
get a correct project and to have the authentication ability
and future functions that is built in django user model
here are few things:
- I have added a one to one field for each user and supplier
that is pointed to a user account in django user model
- I have changed the migration tests in order to get
a correct test that do not fail
- Updated some of few functions in client and supplier
- Update confest file in order to get rid off warnings
- Updated pytest test files to have a correct tests

nothing